geo-tests-v09
Improve test coverage of geometric types

The tests for the geometric functions and operators were in sad state.
This commit increases test coverage of geo_ops.c significantly.  It
removes the alternative results to expect -0 on some platforms.  We
better try to return the same results on different platforms, but if we
can't we can add them back using the results from build farm.

The tests are added to geometric.sql file.  It is not clear where to
put them as there are many cross datatype operators.  Organising them
on a single file sorted by the left hand side of the operators appear
to be a simple solution.  We may take this further and remove the other
tests later.

Discussion: https://www.postgresql.org/message-id/CAE2gYzxF7-5djV6-cEvqQu-fNsnt%3DEqbOURx7ZDg%2BVv6ZMTWbg%40mail.gmail.com