geo-float-v14
Use the built-in float datatype to implement geometric types

This patch makes the geometric operators and functions use the exported
function of the float datatype.  The main reason of doing so is to check
for underflow and overflow, and to handle NaNs consciously.

The float datatypes consider NaNs to be equal and greater than any
non-NaN.  This change considers NaNs equal only for the equality
operators.  The placement operators, contains, overlaps, left/right of
etc. continues to return false when NaNs are involved.  We don't need
to worry about them being considered greater than any-NaN because there
aren't any basic comparison operators like less/greater than for the
geometric datatypes.

All changes can be summarised as:

* Check for underflow and overflow
* Check for division by zero
* Let the objects with NaN values to be the same
* Return NULL when the distance is NaN for all closest point operators
* Favour not-NaN over NaN where it makes sense

The patch also replaces all occurrences of "double" as "float8".  They
are the same, but were randomly spread on the same file.

Discussion:
https://www.postgresql.org/message-id/CAE2gYzxF7-5djV6-cEvqQu-fNsnt%3DEqbOURx7ZDg%2BVv6ZMTWbg%40mail.gmail.com